added file to Code:Blocks project

+++ src/gui/Main.cpp
@@ -906,7 +906,7 @@ void FunctionWindow::OnTestSaveFile(wxCommandEvent& event)
     std::cerr << "Pressed Test Save File" << std::endl;
     sData TestObjekt;
     sData TestObjekt2;
-    
+
     int xpt = 1;
     int ypt = 1;
     int count = 10000;
@@ -918,8 +918,8 @@ void FunctionWindow::OnTestSaveFile(wxCommandEvent& event)
     TestObjekt.setEndFreq(endFreq);
 
     std::vector<double> TestArray;
-    
-    for (size_t i = 0; i < count; i++)
+
+    for (int i = 0; i < count; i++)
     {
         try
         {
@@ -928,32 +928,32 @@ void FunctionWindow::OnTestSaveFile(wxCommandEvent& event)
         catch(const std::exception& e)
         {
             std::cerr << "pushback failed" << e.what() << '\n';
-        } 
+        }
     }
     std::cout << "TestArray Ok" << std::endl;
     std::cout << "count: " << count << std::endl;
     std::cout << "real size :" << TestArray.size() << std::endl;
     TestObjekt.setNumberofPts_Array(count);
     std::cout << "setNumberofPts_Array Ok: " << TestObjekt.getNumberOfPts_Array()<< std::endl;
-    for (size_t i = 0; i < xpt; i++)
+    for (int i = 0; i < xpt; i++)
     {
-        for (size_t j = 0; j < ypt; j++)
+        for (int j = 0; j < ypt; j++)
         {
             try
             {
                 std::cout << "x: " << i << " y: " << j << std::endl;
-                
+
                 TestObjekt.set3DDataReal(TestArray,i,j);
             }
             catch(const std::exception& e)
             {
                 std::cerr << "Set 3D Data failed: " << e.what() << '\n';
                 std::cerr << i << " " << j << std::endl;
-            }  
+            }
         }
     }
     std::cout << "Set 3D Data Ok" << std::endl;
-    
+
 
 
     sData::sParam* TestData = TestObjekt.GetParameter();
